requirements.txt
nltk>=3.8
spacy>=3.4.0
scikit-learn>=1.0.0
pandas>=1.3.0
transformers>=4.20.0
torch>=1.12.0
numpy>=1.21.0
langchain-community>=0.3.0
langchain-huggingface>=1.0.0
sentence-transformers>=2.0.0
langchain-google-genai>=2.0.0
faiss-cpu>=1.7.0
flair>=0.12.0
tavily-python